Move styleguide theme out of the page folder.
